Voice activity detection
Detect when the user is done speaking, and trigger the
response reflex.

whisper.cpp's VAD behaves just like VAD in most video calling
applications. This can cause false positives due to noise.
Added an additional filter which checks if the user recently
said words.

This depends on a patch to whisper.cpp