Display a graph of the conversation
Events are used to create a simple graphviz dot file:

  A [label: "my event"]
  A -> B

Event types and timestamps are used to create unique event IDs.
We do not support every graphviz feature, only the most basic
labels and connections.

Run "npm run graph" to create an SVG and open it. This requires
that graphviz's "dot" program be installed.